% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{filter_country}
\alias{filter_country}
\title{Filter Data to a Single Country}
\usage{
filter_country(data, country)
}
\arguments{
\item{data}{A data frame containing a column called country_name.}

\item{country}{A string with the name of the country to filter to.}
}
\value{
A filtered data frame.
}
\description{
A helper function that filters a data frame to rows matching a given country.
}
