% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_data}
\alias{load_data}
\title{Load Democracy Data}
\usage{
load_data()
}
\value{
A data frame containing democracy data from 1950-2022.
}
\description{
Loads the democracy dataset from the TidyTuesday repository.
}
