% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-regime-changes.R
\name{plot_regime_changes}
\alias{plot_regime_changes}
\title{Plot Regime Changes by Country}
\usage{
plot_regime_changes(data, top_n = 10, color_by = "continent")
}
\arguments{
\item{data}{A data frame containing democracy data with columns
country_name, regime_category_index, and year.}

\item{top_n}{Number of countries to display. Default is 10.}

\item{color_by}{Column to color bars by. Default is "continent".}
}
\value{
A ggplot object.
}
\description{
Creates a horizontal bar chart of the most politically unstable countries.
}
